Several minor changes
- Move some schema conversion from fetchSchema to convertSchema
- Use column.typeOptions.foreignTableId to lookup table name
- Add start command to cli for even easier server launch
