+++ bin/cli.js
@@ -1,8 +1,9 @@
 #! /usr/bin/env node
 
-const fs = require('fs')
+const fs = require('fs');
 const fetchSchema = require('../fetchSchema');
-var program = require('commander');
+const AirtableGraphQL = require('../index');
+const program = require('commander');
 
 program
   .command('pull')
@@ -17,6 +18,15 @@ program
     }).then(schema => {
       fs.writeFileSync('./schema.json', JSON.stringify(schema, null, 2), 'utf-8'); 
     })
-  })
+  });
+
+program
+  .command('start')
+  .option('-s --schema [path]', 'Path of the file containing Airtable schema', './schema.json')
+  .option('-p --port [port]', 'Port for the adapter to listen on', '8765')
+  .action(function (cmd) {
+    const api = new AirtableGraphQL(process.env.AIRTABLE_API_KEY, {schemaPath: cmd.schema});
+    api.listen({port: cmd.port}).then(() => console.log('done')).catch((e) => console.log(e));
+  });
 
 program.parse(process.argv);
